chap01/w1/src/main/java/org/fp024/w1/MyServlet.java
package org.fp024.w1;

import jakarta.servlet.annotation.WebServlet;
import jakarta.servlet.http.HttpServlet;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import java.io.IOException;
import java.io.PrintWriter;

@WebServlet(name = "myServlet", value = "/my")
public class MyServlet extends HttpServlet {

  @Override
  protected void doGet(HttpServletRequest request, HttpServletResponse response)
      throws IOException {
    response.setContentType("text/html");

    PrintWriter out = response.getWriter();

    out.print(
        """
        <html>
        <body>
          <h1>MyServlet</h1>
        </body>
        </html>
        """);
  }
}
